Separated API into different class
